Clean up into utils
